feat: adicionado um exception e handler para que possamos ver como e aonde implementar na arquitetura.

+++ src/main/java/io/meuprojeto/clean_arch_demo/usecase/deleteUser/DeleteUserUseCase.java
@@ -1,6 +1,7 @@
 package io.meuprojeto.clean_arch_demo.usecase.deleteUser;
 
 import io.meuprojeto.clean_arch_demo.domain.repository.UserRepository;
+import io.meuprojeto.clean_arch_demo.usecase.exceptions.NotExistingUserException;
 
 public class DeleteUserUseCase {
 
@@ -11,7 +12,10 @@ public class DeleteUserUseCase {
     }
 
     public String execute(Long id) {
+        if (!repository.findById(id).isPresent()) {
+            throw new NotExistingUserException("User with ID (" + id + ") does not exist.");
+        }
         repository.deleteById(id);
-        return "User with ID " + id + " has been deleted.";
+        return "User with ID (" + id + ") has been deleted.";
     }
 }
